+++ .dockerignore
@@ -234,7 +234,7 @@ cover/
 # Django stuff:
 *.log
 local_settings.py
-db.sqlite3
+data/db.sqlite3
 db.sqlite3-journal
 
 # Flask stuff:
