+++ .gitignore
@@ -58,7 +58,7 @@ coverage.xml
 # Django stuff:
 *.log
 local_settings.py
-db.sqlite3
+data/db.sqlite3
 db.sqlite3-journal
 
 # Flask stuff:
