+++ src/models/__init__.py
@@ -40,6 +40,4 @@ def init_db():
 
 
 # PEP8 says these shouldn't be here, but putting these here avoids circular references within this module.
-from models.registration import Registration
-from models.convostate import ConvoState
-from models.participant import Participant
+from src.models.registration import Registration
